requirements.txt
langchain
langchain-openai>=0.0.5
langchain-neo4j
langchain-community
langchain-chroma
neo4j
chromadb
wikipedia-api
python-dotenv
streamlit>=1.32.0
pandas
openpyxl 
plotly